filter api implemented.

+++ lib/model/request_models/item_filter_request_model.dart
@@ -1,7 +1,7 @@
 class ItemsFilterRequestModel {
   List<String>? cuisineType;
   PriceRange? priceRange;
-  int? minRating;
+  double? minRating;
 
   ItemsFilterRequestModel({this.cuisineType, this.priceRange, this.minRating});
